Claude/blender GitHub pages tht l7

+++ source/blender/gpu/opengl/gl_index_buffer.cc
@@ -70,7 +70,7 @@ void GLIndexBuf::bind_as_ssbo(uint binding)
 
 #ifndef NDEBUG
   BLI_assert(binding < 16);
-  GLContext::get()->bound_ssbo_slots |= 1 << binding;
+  GLContext::get()->bound_ssbo_slots |= uint16_t(1) << binding;
 #endif
 }
